Adding fiat option if fiat funding source is available
I think prob best play if its available is to do
https://stripe-payment-link.com?lightning=lnbc..

that way the invoice could be scanned with a qrcode scanner to pay the stripe invoice and then also scanned with bitcoin wallet to pay.
Useful for selling OTC bits, and for merchants who don't want to be running 2 pos solutions.